No account

+++ src/front/js/store/flux.js
@@ -1,10 +1,12 @@
 const getState = ({ getStore, getActions, setStore }) => {
 	return {
 		store: {
-			userToDos: {},
-			userContacts: {},
-			userMemos: {},
-			userName: []
+			userToDos: [],
+			userContacts: [],
+			userMemos: [],
+			userName: [],
+			loggedIn: false,
+			idCounter: 0
 		},
 		actions: {
 			fetchAPI: async (url, method, body, app) => {
@@ -21,6 +23,7 @@ const getState = ({ getStore, getActions, setStore }) => {
 					const data = await response.json();
 					if (response.status === 200 && app === "ToDos"){
 						setStore({userToDos: data.todos});
+						console.log(getStore().userToDos)
 						return true;
 					};
 					if (response.status === 200 && app === "Contacts"){
@@ -34,6 +37,8 @@ const getState = ({ getStore, getActions, setStore }) => {
 					if (response.status === 200 && app === "signIn"){
 						sessionStorage.setItem('token', data.token);
 						setStore({userName: data.name});
+						setStore({loggedIn: true});
+						console.log(getStore().loggedIn)
 						return true;
 					}
 				} catch (error) {
@@ -41,28 +46,48 @@ const getState = ({ getStore, getActions, setStore }) => {
 				}
 			},
 			fetchUserToDos: async () => {
-				const url = `${process.env.BACKEND_URL}/api/todos`;
-				const method = 'GET';
-				const body = undefined;
-				const app = "ToDos"
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/todos`;
+					const method = 'GET';
+					const body = undefined;
+					const app = "ToDos"
+					getActions().fetchAPI(url, method, body, app)
+				}
 			},
 			addUserToDos: async (task) =>{
-				const url = `${process.env.BACKEND_URL}/api/todos`;
-				const method = 'POST';
-				const body = JSON.stringify({
-					'task': task,
-					'status': 'Not Started'
-				});
-				const app = "ToDos"
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/todos`;
+					const method = 'POST';
+					const body = JSON.stringify({
+						'task': task,
+						'status': 'Not Started'
+					});
+					const app = "ToDos"
+					getActions().fetchAPI(url, method, body, app)
+				} else {
+					const currentToDos = getStore().userToDos;
+					const newToDo = {
+						task: task,
+						status: 'Not Started',
+						id: getStore().idCounter
+					};
+					let newToDoList = currentToDos.concat(newToDo);
+					setStore({ userToDos: newToDoList});
+					setStore({ idCounter: getStore().idCounter + 1});
+				}
 			},
 			deleteUserToDos: async(todoID) =>{
-				const url = `${process.env.BACKEND_URL}/api/todos/${todoID}`;
-				const method = 'DELETE';
-				const body = undefined;
-				const app = "ToDos"
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/todos/${todoID}`;
+					const method = 'DELETE';
+					const body = undefined;
+					const app = "ToDos"
+					getActions().fetchAPI(url, method, body, app)
+				} else {
+					let newToDoList = getStore().userToDos.filter((item) => item.id != todoID);
+					setStore({ userToDos: newToDoList})
+					console.log(getStore().userToDos)
+				}
 			},
 			editToDoStatus: async (todoID, status) => {
 				const opts = {
@@ -90,23 +115,39 @@ const getState = ({ getStore, getActions, setStore }) => {
 				}
 			},
 			fetchUserContacts: async () => {
-				const url = `${process.env.BACKEND_URL}/api/contacts`;
-				const method = 'GET';
-				const body = undefined;
-				const app = "Contacts";
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/contacts`;
+					const method = 'GET';
+					const body = undefined;
+					const app = "Contacts";
+					getActions().fetchAPI(url, method, body, app)
+				}
 			},
 			addUserContact: async (contact) => {
-				const url = `${process.env.BACKEND_URL}/api/contacts`;
-				const method = 'POST';
-				const body = JSON.stringify({
-					'name': contact.name,
-					'phone': contact.phone,
-					'email': contact.email,
-					'address': contact.address
-				});
-				const app = "Contacts";
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/contacts`;
+					const method = 'POST';
+					const body = JSON.stringify({
+						'name': contact.name,
+						'phone': contact.phone,
+						'email': contact.email,
+						'address': contact.address
+					});
+					const app = "Contacts";
+					getActions().fetchAPI(url, method, body, app)
+				} else {
+					const currentContacts = getStore().userContacts;
+					const newContact = {
+						name: contact.name,
+						phone: contact.phone,
+						email: contact.email,
+						address: contact.address,
+						id: getStore().idCounter
+					};
+					let newContactList = currentContacts.concat(newContact);
+					setStore({ userContacts: newContactList});
+					setStore({ idCounter: getStore().idCounter + 1});
+				}
 			},
 			editUserContact: async (contact, contactID) => {
 				const url = `${process.env.BACKEND_URL}/api/contacts/${contactID}`;
@@ -121,28 +162,48 @@ const getState = ({ getStore, getActions, setStore }) => {
 				getActions().fetchAPI(url, method, body, app)
 			},
 			deleteUserContact: async (contactID) => {
-				const url = `${process.env.BACKEND_URL}/api/contacts/${contactID}`;
-				const method = 'DELETE';
-				const body = undefined;
-				const app = "Contacts";
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/contacts/${contactID}`;
+					const method = 'DELETE';
+					const body = undefined;
+					const app = "Contacts";
+					getActions().fetchAPI(url, method, body, app)
+				} else {
+					let newContactList = getStore().userContacts.filter((item) => item.id != contactID);
+					setStore({ userContacts: newContactList})
+					console.log(getStore().userContacts)
+				}
 			},
 			fetchUserMemos: async () => {
-				const url = `${process.env.BACKEND_URL}/api/memos`;
-				const method = 'GET';
-				const body = undefined;
-				const app = "Memos";
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/memos`;
+					const method = 'GET';
+					const body = undefined;
+					const app = "Memos";
+					getActions().fetchAPI(url, method, body, app)
+				}
 			},
 			addUserMemo: async (memo) => {
-				const url = `${process.env.BACKEND_URL}/api/memos`;
-				const method = 'POST';
-				const body = JSON.stringify({
-					'title': memo.title,
-					'memo_body': memo.body
-				});
-				const app = "Memos";
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/memos`;
+					const method = 'POST';
+					const body = JSON.stringify({
+						'title': memo.title,
+						'memo_body': memo.body
+					});
+					const app = "Memos";
+					getActions().fetchAPI(url, method, body, app)
+				} else {
+					const currentMemos = getStore().userMemos;
+					const newMemo = {
+						title: memo.title,
+						memo_body: memo.body,
+						id: getStore().idCounter
+					};
+					const newMemoList = currentMemos.concat(newMemo);
+					setStore({ userMemos: newMemoList});
+					setStore({ idCounter: getStore().idCounter + 1});
+				}
 			},
 			editUserMemo: async (memo) => {
 				const url = `${process.env.BACKEND_URL}/api/memos/${memo.id}`;
@@ -155,11 +216,17 @@ const getState = ({ getStore, getActions, setStore }) => {
 				getActions().fetchAPI(url, method, body, app)
 			},
 			deleteUserMemo: async (memoID) => {
-				const url = `${process.env.BACKEND_URL}/api/memos/${memoID}`;
-				const method = 'DELETE';
-				const body = undefined;
-				const app = "Memos";
-				getActions().fetchAPI(url, method, body, app)
+				if(getStore().loggedIn == true){
+					const url = `${process.env.BACKEND_URL}/api/memos/${memoID}`;
+					const method = 'DELETE';
+					const body = undefined;
+					const app = "Memos";
+					getActions().fetchAPI(url, method, body, app)
+				} else {
+					let newMemoList = getStore().userMemos.filter((item) => item.id != memoID);
+					setStore({ userMemos: newMemoList})
+					console.log(getStore().userMemos)
+				}
 			},
 			createUser: async (user) => {
 				const url = `${process.env.BACKEND_URL}/api/user/new`;
@@ -170,7 +237,13 @@ const getState = ({ getStore, getActions, setStore }) => {
 					'password': user.password
 				});
 				const app = undefined;
-				getActions().fetchAPI(url, method, body, app)
+				getActions().fetchAPI(url, method, body, app);
+				// if (getStore().userToDos.length > 0){
+				// 	const userToDos = getStore().userToDos;
+				// 	for (task in userToDos) {
+
+				// 	}
+				// }
 			},
 			userSignIn: async (user) => {
 				const url = `${process.env.BACKEND_URL}/api/user`;
@@ -184,6 +257,7 @@ const getState = ({ getStore, getActions, setStore }) => {
 			},
 			userLogOut: () => {
 				sessionStorage.removeItem('token');
+				setStore({loggedIn: false});
 			}
 		}
 	};
